fix: add submission handling to submissions controller
